Numerous changes.
Added cable routing channels on the back corner pieces, that are then clamped down/hidden by the back panel.

Adjusted the A and B motor mounts to be able to slide back and forth so that they can be independently tensioned correctly.

Added the beginnings of hinges to the front of the printer on the top side.