So much stuff.
Made panel designs for the enclosure.
Moved away from having the weird stub, so that I could later put a magnet on the door.

Changed the design of the right y carriage to make it possible to get the full travel over the bed.

Made changes to the back center pieces on the top and bottom to allow for cable routing.